# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"CreateTaskerRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"specialties": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"phone",
					"specialties"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTaskerRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"specialties": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"phone",
					"specialties"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"_36_Enums.InfoCategory": {
				"type": "string",
				"enum": [
					"JOB",
					"SCHOLARSHIP",
					"COMPETITION",
					"COMMUNITY",
					"ADVISORY"
				]
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rise-motive-be",
		"version": "1.0.0",
		"description": "Rise Motive Backend API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/taskers": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"properties": {
													"isActive": {
														"type": "boolean"
													},
													"specialties": {
														"type": "string"
													},
													"email": {
														"type": "string"
													},
													"phone": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													}
												},
												"required": [
													"isActive",
													"specialties",
													"email",
													"phone",
													"name",
													"id",
													"createdAt"
												],
												"type": "object"
											},
											"type": "array"
										},
										"total": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"items",
										"total"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Taskers"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isActive": {
											"type": "boolean"
										},
										"specialties": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"phone": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"isActive",
										"specialties",
										"email",
										"phone",
										"name",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Taskers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskerRequest"
							}
						}
					}
				}
			}
		},
		"/taskers/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isActive": {
											"type": "boolean"
										},
										"specialties": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"phone": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"isActive",
										"specialties",
										"email",
										"phone",
										"name",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Taskers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isActive": {
											"type": "boolean"
										},
										"specialties": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"phone": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"isActive",
										"specialties",
										"email",
										"phone",
										"name",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Taskers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTaskerRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isActive": {
											"type": "boolean"
										},
										"specialties": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"phone": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"isActive",
										"specialties",
										"email",
										"phone",
										"name",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Taskers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/taskers/{id}/toggle-active": {
			"put": {
				"operationId": "ToggleActive",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isActive": {
											"type": "boolean"
										},
										"specialties": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"phone": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"isActive",
										"specialties",
										"email",
										"phone",
										"name",
										"id",
										"createdAt"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Taskers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/info-posts/{id}/toggle-active": {
			"put": {
				"operationId": "ToggleActive",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"contactInfo": {
											"type": "string"
										},
										"applyLink": {
											"type": "string"
										},
										"location": {
											"type": "string"
										},
										"deadline": {
											"type": "string"
										},
										"category": {
											"$ref": "#/components/schemas/_36_Enums.InfoCategory"
										},
										"title": {
											"type": "string"
										},
										"isActive": {
											"type": "boolean"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"description": {
											"type": "string"
										}
									},
									"required": [
										"contactInfo",
										"applyLink",
										"location",
										"deadline",
										"category",
										"title",
										"isActive",
										"id",
										"createdAt",
										"description"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Info Posts"
				],
				"security": [
					{
						"jwt": [
							"SUPER_ADMIN",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}